requirements.txt
fastmcp
google-genai
google-adk
python-dotenv
httpx
uvicorn
litellm
pyyaml
itsdangerous
jinja2
